refactor(cli): split cli.ts into per-command modules + presenters (closes #162)
src/cli.ts (1,981 lines) housed all 12 commands' commander wiring, their
text/json presenters, and the runCli test harness in one file. Split into:

- src/commands/*.ts — one registerXCommand(program) module per command
- src/commands/presenters/*.ts — pure formatting functions (warnings,
  impact, coverage, check, integrate, rename), extracted from nested
  closures so they're independently testable
- src/commands/shared.ts — helpers used by 2+ commands (pathsToEntries,
  applyMode, resolveTestResults, loadIntegrate, parseAgentsFlag)
- src/build-program.ts — composition root wiring all 12 commands onto a
  fresh commander tree
- src/testing/run-cli.ts — the runCli in-process test harness, moved out
  of the production composition root
- src/hook-stdin-override.ts — the hook-pretool stdin test seam, shared
  between the harness and the hook-pretool command without a cli.ts
  dependency in either direction

src/cli.ts is now a thin entry point: builds the program via
buildProgram(), re-exports runCli for the test seam tests import from
../src/cli.js, and keeps the process.argv entry-point guard.

process.exit / exit-code mapping stays centralized in each command's own
module per the CLI-layer convention; no command logic changed.

Two existing tests grepped src/cli.ts's source text directly to assert
structure (docs-trio-consistency, check-gate-no-regression); updated only
the file path they read from (src/commands/init.ts, src/commands/check.ts)
to match the new layout — same assertions, same intent.

Verified: typecheck, full unit suite (1512 tests), e2e suite (36 tests),
perf suite, and knip all pass with no behavior change.
