+++ src/commands/doctor.ts
@@ -1,0 +1,48 @@
+// `artgraph doctor` — extracted verbatim from `src/cli.ts` (issue #162).
+
+import { Command, Option } from "commander";
+import type { AgentId } from "../agents/descriptors.js";
+import { parseAgentsFlag } from "./shared.js";
+
+// spec 013 T028 — `artgraph doctor` subcommand. Diagnoses Tier 1 distribution
+// health (Skills sha256 / AGENTS.md marker / wrappers / extraneous files).
+// Independent of `artgraph check` per FR-012: the doctor MUST NOT participate
+// in the `check --gate` decision (regression-tested in
+// `tests/check-gate-no-regression.test.ts`).
+// @impl 013-cross-agent-extensions/FR-012
+export function registerDoctorCommand(program: Command): void {
+  program
+    .command("doctor")
+    .description("Diagnose Tier 1 cross-agent distribution health")
+    .option("--agents <list>", "Comma-separated agent ids to diagnose (default: all detected)")
+    .addOption(
+      new Option("--format <format>", "Output format").choices(["text", "json"]).default("text"),
+    )
+    .action(async (opts) => {
+      const rootDir = process.cwd();
+      const { runDoctor, formatDoctorReportJson, formatDoctorReportText } =
+        await import("../doctor.js");
+      // E-adj-A5: parseAgentsFlag centralizes the parseAgentsList catch — same
+      // as init's --agents branch. `init` and `doctor` used to inline a byte-
+      // identical try/catch; migrating doctor onto the helper keeps them in
+      // sync when the error-to-exit behavior changes.
+      const agents: AgentId[] | undefined =
+        opts.agents !== undefined ? parseAgentsFlag(String(opts.agents)) : undefined;
+      // C1 — mirror the `init` action's try/catch so `SkillsInstallError`,
+      // `EACCES` reads, unknown-agent throws, etc. surface as a single
+      // `Error: <msg>` line rather than a raw Node stack trace.
+      try {
+        const report = runDoctor({ rootDir, agents });
+        const out =
+          opts.format === "json" ? formatDoctorReportJson(report) : formatDoctorReportText(report);
+        console.log(out);
+        if (report.summary.failCount > 0) {
+          process.exitCode = 1;
+        }
+      } catch (e) {
+        const msg = e instanceof Error ? e.message : String(e);
+        console.error(`Error: ${msg}`);
+        process.exit(1);
+      }
+    });
+}
